perf(cli): async handlers for deploy commands
