+++ packages/cli/src/commands/deploy/netlify.js
@@ -1,18 +1,23 @@
 import { recordTelemetryAttributes } from '@redmix/cli-helpers'
 
-import { deployBuilder, deployHandler } from './helpers/helpers.js'
+import { deployBuilder } from './helpers/deployBuilder.js'
 
 export const command = 'netlify [...commands]'
 export const description = 'Build command for Netlify deploy'
 
 export const builder = (yargs) => deployBuilder(yargs)
 
-export const handler = (yargs) => {
+export async function handler(yargs) {
   recordTelemetryAttributes({
     command: 'deploy netlify',
     build: yargs.build,
     prisma: yargs.prisma,
     dataMigrate: yargs.dataMigrate,
   })
-  deployHandler(yargs)
+
+  const { handler: importedHandler } = await import(
+    './helpers/deployHandler.js'
+  )
+
+  return importedHandler(yargs)
 }
